requirements.txt
mysqlclient==1.3.12
pkg-resources==0.0.0
tqdm==4.19.8
